agentic_event_orchestrator/requirements.txt
openai
pydantic
python-dotenv
requests
chainlit
ortools
google-generativeai
mcp
